add : color type
