+++ src/components/Typo/Typo.tsx
@@ -5,15 +5,8 @@ import React, {
   ReactNode,
 } from "react";
 import styled from "styled-components";
-import { grayScale, yellow, blue, red, orange, green } from "../../colors";
-
-export type ColorType =
-  | (typeof grayScale)[keyof typeof grayScale]
-  | (typeof yellow)[keyof typeof yellow]
-  | (typeof blue)[keyof typeof blue]
-  | (typeof red)[keyof typeof red]
-  | (typeof orange)[keyof typeof orange]
-  | (typeof green)[keyof typeof green];
+import { grayScale } from "../../colors";
+import { ColorType } from "../../colors/types";
 
 export interface HeadingProps extends HTMLAttributes<HTMLHeadingElement> {
   children: ReactNode | string;
